Move RangeNode to separate class
