+++ Source/DafnyCore/Generic/Node.cs
@@ -201,21 +201,3 @@ public abstract class Node : INode {
     return trivia is not ("" or null);
   }
 }
-
-public abstract class RangeNode : Node { // TODO merge into Node when TokenNode is gone.
-
-  public override IOrigin Tok => StartToken; // TODO rename to ReportingToken in separate PR
-
-  public IOrigin tok => Tok; // TODO replace with Tok in separate PR
-
-  // TODO rename to Range in separate PR
-  public override RangeToken Origin { get; set; } // TODO remove setter when TokenNode is gone.
-
-  protected RangeNode(Cloner cloner, RangeNode original) {
-    Origin = cloner.Tok(original.Origin);
-  }
-
-  protected RangeNode(RangeToken rangeOrigin) {
-    Origin = rangeOrigin;
-  }
-}
